refactor: move object of symbols to script.js
