--- translate.js
@@ -1,144 +1,0 @@
-const i18n = {
-  en: {
-    backquote: '`',
-    backquoteShifted: '~',
-    one: '1',
-    oneShifted: '!',
-    two: '2',
-    twoShifted: '@',
-    three: '3',
-    threeShifted: '#',
-    four: '4',
-    fourShifted: '$',
-    five: '5',
-    fiveShifted: '%',
-    six: '6',
-    sixShifted: '^',
-    seven: '7',
-    sevenShifted: '&',
-    eight: '8',
-    eightShifted: '*',
-    nine: '9',
-    nineShifted: '(',
-    zero: '0',
-    zeroShifted: ')',
-    minus: '-',
-    minusShifted: '_',
-    equal: '=',
-    equalShifted: '+',
-    KeyQ: 'q',
-    KeyW: 'w',
-    KeyE: 'e',
-    KeyR: 'r',
-    KeyT: 't',
-    KeyY: 'y',
-    KeyU: 'u',
-    KeyI: 'i',
-    KeyO: 'o',
-    KeyP: 'p',
-    bracketLeft: '[',
-    bracketLeftShifted: '{',
-    bracketRight: ']',
-    bracketRightShifted: '}',
-    backslash: '\\',
-    backslashShifted: '|',
-    KeyA: 'a',
-    KeyS: 's',
-    KeyD: 'd',
-    KeyF: 'f',
-    KeyG: 'g',
-    KeyH: 'h',
-    KeyJ: 'j',
-    KeyK: 'k',
-    KeyL: 'l',
-    semicolon: ';',
-    semicolonShifted: ':',
-    quote: '\'',
-    quoteShifted: '"',
-    KeyZ: 'z',
-    KeyX: 'x',
-    KeyC: 'c',
-    KeyV: 'v',
-    KeyB: 'b',
-    KeyN: 'n',
-    KeyM: 'm',
-    comma: ',',
-    commaShifted: '<',
-    period: '.',
-    periodShifted: '>',
-    slash: '/',
-    slashShifted: '?',
-  },
-  ru: {
-    backquote: 'ё',
-    backquoteShifted: 'Ё',
-    one: '1',
-    oneShifted: '!',
-    two: '2',
-    twoShifted: '"',
-    three: '3',
-    threeShifted: '№',
-    four: '4',
-    fourShifted: ';',
-    five: '5',
-    fiveShifted: '%',
-    six: '6',
-    sixShifted: ':',
-    seven: '7',
-    sevenShifted: '?',
-    eight: '8',
-    eightShifted: '*',
-    nine: '9',
-    nineShifted: '(',
-    zero: '0',
-    zeroShifted: ')',
-    minus: '-',
-    minusShifted: '_',
-    equal: '=',
-    equalShifted: '+',
-    KeyQ: 'й',
-    KeyW: 'ц',
-    KeyE: 'у',
-    KeyR: 'к',
-    KeyT: 'е',
-    KeyY: 'н',
-    KeyU: 'г',
-    KeyI: 'ш',
-    KeyO: 'щ',
-    KeyP: 'з',
-    bracketLeft: 'х',
-    bracketLeftShifted: 'Х',
-    bracketRight: 'ъ',
-    bracketRightShifted: 'Ъ',
-    backslash: '\\',
-    backslashShifted: '|',
-    KeyA: 'ф',
-    KeyS: 'ы',
-    KeyD: 'в',
-    KeyF: 'а',
-    KeyG: 'п',
-    KeyH: 'р',
-    KeyJ: 'о',
-    KeyK: 'л',
-    KeyL: 'д',
-    semicolon: 'ж',
-    semicolonShifted: 'Ж',
-    quote: 'э',
-    quoteShifted: 'Э',
-    KeyZ: 'я',
-    KeyX: 'ч',
-    KeyC: 'с',
-    KeyV: 'м',
-    KeyB: 'и',
-    KeyN: 'т',
-    KeyM: 'ь',
-    comma: 'б',
-    commaShifted: 'Б',
-    period: 'ю',
-    periodShifted: 'Ю',
-    slash: '.',
-    slashShifted: ',',
-  },
-};
-
-export default i18n;
